+++ assets/js/containers/Application/Application.js
@@ -3,6 +3,7 @@ import { BrowserRouter as Router, Route, Switch } from 'react-router-dom';
 import './global.scss';
 import Login from '../../views/Login';
 import Dashboard from '../../views/Dashboard';
+import TimeTracking from '../../views/TimeTracking';
 import NotFound from '../../views/NotFound';
 import Navigation from '../../components/Navigation';
 
@@ -14,26 +15,27 @@ export default class Application extends React.Component {
     }
 
     render() {
-        if (!this.isLoggedIn()) {
-            return <Login />;
-        }
-
         return (
             <Router>
                 <div>
-                    <Navigation>
-                        <Navigation.Item to="/">
-                            Dashboard
-                        </Navigation.Item>
+                    {
+                        this.isLoggedIn() ? (
+                            <Navigation>
+                                <Navigation.Item to="/">
+                                    Dashboard
+                                </Navigation.Item>
 
-                        <Navigation.Item to="/login">
-                            Login
-                        </Navigation.Item>
-                    </Navigation>
+                                <Navigation.Item to="/timetracking">
+                                    TimeTracking
+                                </Navigation.Item>
+                            </Navigation>
+                        ) : null
+                    }
 
                     <Switch>
                         <Route exact path="/" component={Dashboard} />
                         <Route path="/login" component={Login} />
+                        <Route path="/timetracking" component={TimeTracking} />
                         <Route component={NotFound} />
                     </Switch>
                 </div>