+++ assets/js/views/TimeTracking/index.js
@@ -1,0 +1,3 @@
+import TimeTracking from './TimeTracking';
+
+export default TimeTracking;