+++ CHANGES.md
@@ -2,6 +2,7 @@
 
 ##### Unreleased
 
+- FEATURE: Add security and user entity with example Login view. #41 [!3]
 - FEATURE: Add react-router to frontend setup with example implementation. #41 [!3]
 - FEATURE: Add frontend setup. #41 [!2]
 - FEATURE: Created the Symfony project skeleton. #41 [!1]